Return completion status instead of using callback
