+++ sqlmesh/core/snapshot/evaluator.py
@@ -51,7 +51,7 @@ from sqlmesh.core.model import (
     ViewKind,
     CustomKind,
 )
-
+from sqlmesh.utils import CompletionStatus
 from sqlmesh.core.schema_diff import has_drop_alteration, get_dropped_column_names
 from sqlmesh.core.snapshot import (
     DeployabilityIndex,
@@ -287,10 +287,9 @@ class SnapshotEvaluator:
         snapshots: t.Dict[SnapshotId, Snapshot],
         deployability_index: t.Optional[DeployabilityIndex] = None,
         on_start: t.Optional[t.Callable] = None,
-        on_no_work: t.Optional[t.Callable] = None,
         on_complete: t.Optional[t.Callable[[SnapshotInfoLike], None]] = None,
         allow_destructive_snapshots: t.Optional[t.Set[str]] = None,
-    ) -> None:
+    ) -> CompletionStatus:
         """Creates a physical snapshot schema and table for the given collection of snapshots.
 
         Args:
@@ -298,9 +297,11 @@ class SnapshotEvaluator:
             snapshots: Mapping of snapshot ID to snapshot.
             deployability_index: Determines snapshots that are deployable in the context of this creation.
             on_start: A callback to initialize the snapshot creation progress bar.
-            on_no_work: A callback to call when no snapshots are to be created.
             on_complete: A callback to call on each successfully created snapshot.
             allow_destructive_snapshots: Set of snapshots that are allowed to have destructive schema changes.
+
+        Returns:
+            CompletionStatus: The status of the creation operation (success, failure, nothing to do).
         """
         snapshots_with_table_names = defaultdict(set)
         tables_by_gateway_and_schema: t.Dict[t.Union[str, None], t.Dict[exp.Table, set[str]]] = (
@@ -368,9 +369,7 @@ class SnapshotEvaluator:
                 target_deployability_flags[snapshot.name].sort()
 
         if not snapshots_to_create:
-            if on_no_work:
-                on_no_work("\n[green]SKIP: No physical layer updates to perform[/green]\n")
-            return
+            return CompletionStatus.NOTHING_TO_DO
         if on_start:
             on_start(len(snapshots_to_create))
 
@@ -385,6 +384,7 @@ class SnapshotEvaluator:
             on_complete=on_complete,
             allow_destructive_snapshots=allow_destructive_snapshots,
         )
+        return CompletionStatus.SUCCESS
 
     def _create_snapshots(
         self,
